Change font and clean up the CSS(s)
Since "Courier New" doesn't render well on the web, moving to a
font-stack is to prefer. As discussed earlier, these fonts will
serve well on all/most platforms. Also added the 'Liberation Mono'
as it is a replacement font for Courier (V2) (ie. Courier New),
and uses the same metrics.

The CSS have also been cleaned up a bit - removed some reduntant
code, etc.

- Replace 'Courier New' with a fontstack, and move all fonts to body {}
- Use 1-liners where they fit to 1 line (< 80), and expand if not
- Use hex shortcodes when possible, like: #333 (#333333), #456 (#445566)

- - -

Use a main CSS file, and put the themes (just colors) in separate
files. In that way, for any future style change to the design/layout
there's only need to update one file, instead of 5+ files. For this
layout/theme-separation - the use of CSS variables is really great.
It's also easier when someone wants to add a theme or use a custom
theme, and maybe gets behind on the updates.

- Add manpageblog.css (main css-file)
- Add corresponding link entries in: page{,_tags}.html
- Use CSS variables for color definitions, and set the fallback values
  to 'Light (default theme)'.
- Add theme: 'Light Blue', just like 'Grey' have a 'Grey Blue' version